Fixed cancel booking & booking model
Added status to the booking model, Changed SetSeatOccupied to SetSeatOccupancy. Added BookingStatus to the DB, Added create booking table instead of printing the bookings in every method it now prints it in the createbookingtable method. Made sure the booking status changes when you cancel it and u cant cancel a cancelled booking

Co-Authored-By: Lucior <183356814+EyeMustache@users.noreply.github.com>