UpdateDB now also removed past flight seats
FlightSeatAcces now deletes past flight seats based off a list of flight ids
PastFlightAccess now has a new name for deleting past flights, it uses a list of flight ids to delete instead of a datetime
PastFlightAccess now fetches a list of flight ids past flights older than a month to delete.
Adjusted the interfaces to accept these changes

Updated UpdateFlightDB to handle this

Added lines to Program so show that the program is updating database to show it is doing something and not freezing.