Modify & Cancel Booking Bugfix
New private method in BookingUI handling everything concerning the displaying of the seatmap and the selection of the seat.

Modified DisplayBookingDetails to now only "send" an email address when booking the flight instead of also doing it any time the method is called upon.

Modified Additional Luggage to now get rid of ugly markup lines and instead display a proper message.

Modified AdditionalLuggage and CancellationInsurance to display extra information.

Modifed AddCouponCode to now keep asking for a coupon code after agreeing that you have a code. It'll keep asking until the code filled in is a valid coupon code or until left empty, causing it to return its respective "answer".

Updated Legenda to now also display the pricing of the seats so that the user knows how much the seat will cost during the booking process.

Modified ModifyBookingPrompt to now displayt the seatmap for selecting a different seat.

Made sure Cancellation and Modification of booking is now properly integrated.

Made sure that cancelled flight are marked as "past flights" causing them to 1. only show up in past bookings & 2. be non-eligible for modification of a flight.